+++ store/src/main/java/com/zimbra/cert/GenerateCSR.java
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: GPL-2.0-only
 
-package com.zimbra.cs.certmanager;
+package com.zimbra.cert;
 
 import com.google.common.base.Strings;
 import com.google.common.net.HostSpecifier;
