+++ store/src/main/java/com/zimbra/cert/UploadProxyCA.java
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: GPL-2.0-only
 
-package com.zimbra.cs.certmanager;
+package com.zimbra.cert;
 
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.soap.CertMgrConstants;
@@ -17,20 +17,20 @@ import com.zimbra.soap.ZimbraSoapContext;
 import java.io.IOException;
 import java.util.Map;
 
-public class UploadDomCert extends AdminDocumentHandler {
+public class UploadProxyCA extends AdminDocumentHandler {
 
     @Override
     public Element handle(Element request, Map<String, Object> context) throws ServiceException {
         ZimbraSoapContext lc = getZimbraSoapContext(context);
-        Element response = lc.createElement(CertMgrConstants.UPLOAD_DOMCERT_RESPONSE);
+        Element response = lc.createElement(CertMgrConstants.UPLOAD_PROXYCA_RESPONSE);
 
         String attachId = null;
         String filename = null;
         Upload up = null ;
 
         try {
-            attachId = request.getAttribute(CertMgrConstants.A_CERT_AID) ;
-            filename = request.getAttribute(CertMgrConstants.A_CERT_NAME) ;
+            attachId = request.getAttribute(CertMgrConstants.A_CERT_AID);
+            filename = request.getAttribute(CertMgrConstants.A_CERT_NAME);
             ZimbraLog.security.debug("Found certificate Filename  = " + filename + "; attid = " + attachId );
 
             up = FileUploadServlet.fetchUpload(lc.getAuthtokenAccountId(), attachId, lc.getAuthToken());
@@ -42,24 +42,6 @@ public class UploadDomCert extends AdminDocumentHandler {
                 response.addAttribute(CertMgrConstants.A_cert_content, new String(blob));
         }catch (IOException ioe) {
             throw ServiceException.FAILURE("Can not get uploaded certificate content", ioe);
-        }finally {
-            FileUploadServlet.deleteUpload(up);
-        }
-
-        try {
-            attachId = request.getAttribute(CertMgrConstants.A_KEY_AID);
-            filename = request.getAttribute(CertMgrConstants.A_KEY_NAME);
-            ZimbraLog.security.debug("Found certificate Filename  = " + filename + "; attid = " + attachId );
-
-            up = FileUploadServlet.fetchUpload(lc.getAuthtokenAccountId(), attachId, lc.getAuthToken());
-            if (up == null)
-                throw ServiceException.FAILURE("Uploaded file " + filename + " with " + attachId + " was not found.", null);
-
-            byte [] blob = ByteUtil.getContent(up.getInputStream(),-1) ;
-            if(blob.length > 0)
-                response.addAttribute(CertMgrConstants.A_key_content, new String(blob));
-        }catch (IOException ioe) {
-            throw ServiceException.FAILURE("Can not get uploaded key content", ioe);
         }finally {
             FileUploadServlet.deleteUpload(up);
         }
