+++ store/src/test/java/com/zimbra/cert/GetDomainCertTest.java
@@ -1,4 +1,4 @@
-package com.zimbra.cs.certmanager;
+package com.zimbra.cert;
 
 import static com.zimbra.common.soap.AdminConstants.A_DOMAIN;
 import static com.zimbra.common.soap.CertMgrConstants.E_ISSUER;
@@ -12,6 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.zimbra.cert.GetDomainCert;
 import com.zimbra.common.account.Key.DomainBy;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.soap.Element;
